requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
numpy==1.26.4
scipy==1.12.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==12.0
redis==5.0.1
python-dotenv==1.0.0
cryptography==41.0.5
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0.post1
python-socketio==5.10.0
alembic==1.13.1
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
pytest-cov==4.1.0 